# Conflicts:
#	src/main/java/betterwithmods/common/blocks/mechanical/BlockMechMachines.java
#	src/main/java/betterwithmods/common/entity/ai/EntityAISearchFood.java
#	src/main/java/betterwithmods/event/BlastingOilEvent.java
#	src/main/java/betterwithmods/module/gameplay/miniblocks/blocks/BlockMini.java
#	src/main/java/betterwithmods/module/hardcore/world/HCBuoy.java
#	src/main/java/betterwithmods/module/industry/Industry.java
#	src/main/java/betterwithmods/module/tweaks/CreeperShearing.java
#	src/main/java/betterwithmods/module/tweaks/MushroomFarming.java
#	src/main/resources/assets/betterwithmods/advancements/vertical_windmill.json

+++ src/main/java/betterwithmods/util/commands/HealthCommand.java
@@ -30,7 +30,7 @@ public class HealthCommand extends CommandBase {
     public void execute(@Nonnull MinecraftServer server, @Nonnull ICommandSender sender, @Nonnull String[] args) {
         if (sender instanceof EntityPlayer) {
             if (args.length > 0) {
-                int health = Integer.parseInt(args[0]);
+                float health = Float.parseFloat(args[0]);
                 ((EntityPlayer) sender).setHealth(health);
             }
         }
